requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
sqlparse==0.4.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
